requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.12
django-cors-headers==2.2.0
django-filter==1.1.0
django-rest-auth==0.9.3
django-rest-swagger==2.1.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==5.1.0
psycopg2-binary==2.7.4
PyJWT==1.6.1
pytz==2018.4
requests==2.18.4
simplejson==3.14.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
django-easy-pdf
xhtml2pdf
reportlab
weasyprint